+++ rs-react-app/src/App.css
@@ -1,7 +1,8 @@
 #root {
+  width: 100vw;
+  height: 100vh;
   max-width: 1280px;
   margin: 0 auto;
-  padding: 2rem;
   text-align: center;
 }
 
@@ -39,4 +40,8 @@
 
 .read-the-docs {
   color: #888;
+}
+
+.not-found-msg {
+  margin-top: 50px;
 }
