+++ rs-react-app/src/components/Controls/Controls.module.css
@@ -1,0 +1,8 @@
+.controls {
+  position: sticky;
+  top: 10px;
+  display: flex;
+  justify-content: end;
+  gap: 5px;
+  margin-right: 10px;
+}
